refactor: Object.freeze() 대신 deepFreeze()로 통일
