+++ src/utils/deepFreeze.ts
@@ -1,0 +1,9 @@
+export function deepFreeze<T>(obj: T): T {
+  Object.getOwnPropertyNames(obj).forEach((prop) => {
+    const value = (obj as any)[prop];
+    if (value && typeof value === "object" && !Object.isFrozen(value)) {
+      deepFreeze(value);
+    }
+  });
+  return Object.freeze(obj);
+};
